Extract commands 'affects', 'mode', and 'toggle' to separate files.
