+++ src/act_informative.cpp
@@ -107,7 +107,6 @@ char *diag_obj_to_char(CharData *i, ObjData *obj, int mode);
 const char *diag_obj_timer(const ObjData *obj);
 char *diag_timer_to_char(const ObjData *obj);
 
-void do_affects(CharData *ch, char *argument, int cmd, int subcmd);
 void do_look(CharData *ch, char *argument, int cmd, int subcmd);
 void do_examine(CharData *ch, char *argument, int cmd, int subcmd);
 void do_gold(CharData *ch, char *argument, int cmd, int subcmd);
@@ -125,7 +124,6 @@ void do_where(CharData *ch, char *argument, int cmd, int subcmd);
 void do_levels(CharData *ch, char *argument, int cmd, int subcmd);
 void do_consider(CharData *ch, char *argument, int cmd, int subcmd);
 void do_diagnose(CharData *ch, char *argument, int cmd, int subcmd);
-void do_toggle(CharData *ch, char *argument, int cmd, int subcmd);
 void SortCommands();
 void do_commands(CharData *ch, char *argument, int cmd, int subcmd);
 void do_looking(CharData *ch, char *argument, int cmd, int subcmd);
@@ -3740,140 +3738,6 @@ void do_diagnose(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
 	}
 }
 
-const char *ctypes[] = {"выключен", "простой", "обычный", "полный", "\n"};
-
-void do_toggle(CharData *ch, char * /*argument*/, int/* cmd*/, int/* subcmd*/) {
-	if (ch->IsNpc())
-		return;
-	if (GET_WIMP_LEV(ch) == 0)
-		strcpy(buf2, "нет");
-	else
-		sprintf(buf2, "%-3d", GET_WIMP_LEV(ch));
-
-	if (GetRealLevel(ch) >= kLvlImmortal || PRF_FLAGGED(ch, EPrf::kCoderinfo)) {
-		snprintf(buf, kMaxStringLength,
-				 " Нет агров     : %-3s     "
-				 " Супервидение  : %-3s     "
-				 " Флаги комнат  : %-3s \r\n"
-				 " Частный режим : %-3s     "
-				 " Замедление    : %-3s     "
-				 " Кодер         : %-3s \r\n"
-				 " Опечатки      : %-3s \r\n",
-				 ONOFF(PRF_FLAGGED(ch, EPrf::kNohassle)),
-				 ONOFF(PRF_FLAGGED(ch, EPrf::kHolylight)),
-				 ONOFF(PRF_FLAGGED(ch, EPrf::kRoomFlags)),
-				 ONOFF(PRF_FLAGGED(ch, EPrf::kNoWiz)),
-				 ONOFF(nameserver_is_slow),
-				 ONOFF(PRF_FLAGGED(ch, EPrf::kCoderinfo)),
-				 ONOFF(PRF_FLAGGED(ch, EPrf::kShowUnread)));
-		SendMsgToChar(buf, ch);
-	}
-
-	snprintf(buf, kMaxStringLength,
-			 " Автовыходы    : %-3s     "
-			 " Краткий режим : %-3s     "
-			 " Сжатый режим  : %-3s \r\n"
-			 " Повтор команд : %-3s     "
-			 " Обращения     : %-3s     "
-			 " Цвет          : %-8s \r\n"
-			 " Кто-то        : %-6s  "
-			 " Болтать       : %-3s     "
-			 " Орать         : %-3s \r\n"
-			 " Аукцион       : %-3s     "
-			 " Базар         : %-3s     "
-			 " Автозаучивание: %-3s \r\n"
-			 " Призыв        : %-3s     "
-			 " Автозавершение: %-3s     "
-			 " Группа (вид)  : %-7s \r\n"
-			 " Без двойников : %-3s     "
-			 " Автопомощь    : %-3s     "
-			 " Автодележ     : %-3s \r\n"
-			 " Автограбеж    : %-7s "
-			 " Брать куны    : %-3s     "
-			 " Арена         : %-3s \r\n"
-			 " Трусость      : %-3s     "
-			 " Ширина экрана : %-3d     "
-			 " Высота экрана : %-3d \r\n"
-			 " Сжатие        : %-6s  "
-			 " Новости (вид) : %-5s   "
-			 " Доски         : %-3s \r\n"
-			 " Хранилище     : %-8s"
-			 " Пклист        : %-3s     "
-			 " Политика      : %-3s \r\n"
-			 " Пкформат      : %-6s  "
-			 " Соклановцы    : %-8s"
-			 " Оффтоп        : %-3s \r\n"
-			 " Потеря связи  : %-3s     "
-			 " Ингредиенты   : %-3s     "
-			 " Вспомнить     : %-3u \r\n",
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kAutoexit)),
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kBrief)),
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kCompact)),
-			 YESNO(!PRF_FLAGGED(ch, EPrf::kNoRepeat)),
-			 ONOFF(!PRF_FLAGGED(ch, EPrf::kNoTell)),
-			 ctypes[COLOR_LEV(ch)],
-			 PRF_FLAGGED(ch, EPrf::kNoInvistell) ? "нельзя" : "можно",
-			 ONOFF(!PRF_FLAGGED(ch, EPrf::kNoGossip)),
-			 ONOFF(!PRF_FLAGGED(ch, EPrf::kNoHoller)),
-			 ONOFF(!PRF_FLAGGED(ch, EPrf::kNoAuction)),
-			 ONOFF(!PRF_FLAGGED(ch, EPrf::kNoExchange)),
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kAutomem)),
-			 ONOFF(PRF_FLAGGED(ch, EPrf::KSummonable)),
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kGoAhead)),
-			 PRF_FLAGGED(ch, EPrf::kShowGroup) ? "полный" : "краткий",
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kNoClones)),
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kAutoassist)),
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kAutosplit)),
-			 PRF_FLAGGED(ch, EPrf::kAutoloot) ? PRF_FLAGGED(ch, EPrf::kNoIngrLoot) ? "NO-INGR" : "ALL    " : "OFF    ",
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kAutomoney)),
-			 ONOFF(!PRF_FLAGGED(ch, EPrf::kNoArena)),
-			 buf2,
-			 STRING_LENGTH(ch),
-			 STRING_WIDTH(ch),
-#if defined(HAVE_ZLIB)
-			 ch->desc->deflate == nullptr ? "нет" : (ch->desc->mccp_version == 2 ? "MCCPv2" : "MCCPv1"),
-#else
-		"N/A",
-#endif
-			 PRF_FLAGGED(ch, EPrf::kNewsMode) ? "доска" : "лента",
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kBoardMode)),
-			 GetChestMode(ch).c_str(),
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kPklMode)),
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kPolitMode)),
-			 PRF_FLAGGED(ch, EPrf::kPkFormatMode) ? "краткий" : "полный",
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kClanmembersMode)),
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kOfftopMode)),
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kAntiDcMode)),
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kNoIngrMode)),
-			 ch->remember_get_num());
-	SendMsgToChar(buf, ch);
-	if (NOTIFY_EXCH_PRICE(ch) > 0) {
-		sprintf(buf, " Уведомления   : %-7ld ", NOTIFY_EXCH_PRICE(ch));
-	} else {
-		sprintf(buf, " Уведомления   : %-7s ", "Нет");
-	}
-	SendMsgToChar(buf, ch);
-	snprintf(buf, kMaxStringLength,
-			 " Карта         : %-3s     "
-			 " Вход в зону   : %-3s   \r\n"
-			 " Магщиты (вид) : %-8s"
-			 " Автопризыв    : %-5s   "
-			 " Маппер        : %-3s   \r\n"
-			 " Контроль IP   : %-6s  ",
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kDrawMap)),
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kShowZoneNameOnEnter)),
-			 (PRF_FLAGGED(ch, EPrf::kBriefShields) ? "краткий" : "полный"),
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kAutonosummon)),
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kMapper)),
-			 ONOFF(PRF_FLAGGED(ch, EPrf::kIpControl)));
-	SendMsgToChar(buf, ch);
-	if (GET_GOD_FLAG(ch, EGf::kAllowTesterMode))
-		sprintf(buf, " Тестер        : %-3s\r\n", ONOFF(PRF_FLAGGED(ch, EPrf::kTester)));
-	else
-		sprintf(buf, "\r\n");
-	SendMsgToChar(buf, ch);
-}
-
 void do_zone(CharData *ch, char * /*argument*/, int/* cmd*/, int/* subcmd*/) {
 	if (ch->desc
 		&& !(is_dark(ch->in_room) && !CAN_SEE_IN_DARK(ch) && !CanUseFeat(ch, EFeat::kDarkReading))
@@ -3977,124 +3841,6 @@ void do_commands(CharData *ch, char *argument, int/* cmd*/, int subcmd) {
 	SendMsgToChar(buf, ch);
 }
 
-std::array<EAffect, 3> hiding = {EAffect::kSneak,
-								 EAffect::kHide,
-								 EAffect::kDisguise};
-
-void do_affects(CharData *ch, char * /*argument*/, int/* cmd*/, int/* subcmd*/) {
-	char sp_name[kMaxStringLength];
-
-	// Show the bitset without "hiding" etc.
-	auto aff_copy = ch->char_specials.saved.affected_by;
-	for (auto j : hiding) {
-		aff_copy.unset(j);
-	}
-
-	aff_copy.sprintbits(affected_bits, buf2, ",");
-	snprintf(buf, kMaxStringLength, "Аффекты: %s%s%s\r\n", CCIYEL(ch, C_NRM), buf2, CCNRM(ch, C_NRM));
-	SendMsgToChar(buf, ch);
-
-	// Routine to show what spells a char is affected by
-	if (!ch->affected.empty()) {
-		for (auto affect_i = ch->affected.begin(); affect_i != ch->affected.end(); ++affect_i) {
-			const auto aff = *affect_i;
-
-			if (aff->type == ESpell::kSolobonus) {
-				continue;
-			}
-
-			*buf2 = '\0';
-			strcpy(sp_name, MUD::Spell(aff->type).GetCName());
-			int mod = 0;
-			if (aff->battleflag == kAfPulsedec) {
-				mod = aff->duration / 51; //если в пульсах приводим к тикам 25.5 в сек 2 минуты
-			} else {
-				mod = aff->duration;
-			}
-			(mod + 1) / kSecsPerMudHour
-			? sprintf(buf2,
-					  "(%d %s)",
-					  (mod + 1) / kSecsPerMudHour + 1,
-					  GetDeclensionInNumber((mod + 1) / kSecsPerMudHour + 1, EWhat::kHour))
-			: sprintf(buf2, "(менее часа)");
-			snprintf(buf, kMaxStringLength, "%s%s%-21s %-12s%s ",
-					 *sp_name == '!' ? "Состояние  : " : "Заклинание : ",
-					 CCICYN(ch, C_NRM), sp_name, buf2, CCNRM(ch, C_NRM));
-			*buf2 = '\0';
-			if (!IS_IMMORTAL(ch)) {
-				auto next_affect_i = affect_i;
-				++next_affect_i;
-				if (next_affect_i != ch->affected.end()) {
-					const auto &next_affect = *next_affect_i;
-					if (aff->type == next_affect->type) {
-						continue;
-					}
-				}
-			} else {
-				if (aff->modifier) {
-					sprintf(buf2, "%-3d к параметру: %s", aff->modifier, apply_types[(int) aff->location]);
-					strcat(buf, buf2);
-				}
-				if (aff->bitvector) {
-					if (*buf2) {
-						strcat(buf, ", устанавливает ");
-					} else {
-						strcat(buf, "устанавливает ");
-					}
-					strcat(buf, CCIRED(ch, C_NRM));
-					sprintbit(aff->bitvector, affected_bits, buf2);
-					strcat(buf, buf2);
-					strcat(buf, CCNRM(ch, C_NRM));
-				}
-			}
-			SendMsgToChar(strcat(buf, "\r\n"), ch);
-		}
-// отображение наград
-		for (const auto &aff : ch->affected) {
-			if (aff->type == ESpell::kSolobonus) {
-				int mod;
-				if (aff->battleflag == kAfPulsedec) {
-					mod = aff->duration / 51; //если в пульсах приводим к тикам	25.5 в сек 2 минуты
-				} else {
-					mod = aff->duration;
-				}
-				(mod + 1) / kSecsPerMudHour
-				? sprintf(buf2,
-						  "(%d %s)",
-						  (mod + 1) / kSecsPerMudHour + 1,
-						  GetDeclensionInNumber((mod + 1) / kSecsPerMudHour + 1, EWhat::kHour))
-				: sprintf(buf2, "(менее часа)");
-				snprintf(buf,
-						 kMaxStringLength,
-						 "Заклинание : %s%-21s %-12s%s ",
-						 CCICYN(ch, C_NRM),
-						 "награда",
-						 buf2,
-						 CCNRM(ch, C_NRM));
-				*buf2 = '\0';
-				if (aff->modifier) {
-					sprintf(buf2, "%s%-3d к параметру: %s%s%s", (aff->modifier > 0) ? "+" : "",
-							aff->modifier, CCIRED(ch, C_NRM), apply_types[(int) aff->location], CCNRM(ch, C_NRM));
-					strcat(buf, buf2);
-				}
-				SendMsgToChar(strcat(buf, "\r\n"), ch);
-			}
-		}
-	}
-
-	if (ch->is_morphed()) {
-		*buf2 = '\0';
-		SendMsgToChar("Автоаффекты звериной формы: ", ch);
-		const IMorph::affects_list_t &affs = ch->GetMorphAffects();
-		for (auto it = affs.cbegin(); it != affs.cend();) {
-			sprintbit(to_underlying(*it), affected_bits, buf2);
-			SendMsgToChar(string(CCIYEL(ch, C_NRM)) + string(buf2) + string(CCNRM(ch, C_NRM)), ch);
-			if (++it != affs.end()) {
-				SendMsgToChar(", ", ch);
-			}
-		}
-	}
-}
 
 // Create web-page with users list
 void make_who2html() {
