+++ src/cmd/do_affects.h
@@ -1,0 +1,14 @@
+//
+// Created by Sventovit on 07.09.2024.
+//
+
+#ifndef BYLINS_SRC_CMD_DO_AFFECTS_H_
+#define BYLINS_SRC_CMD_DO_AFFECTS_H_
+
+class CharData;
+
+void do_affects(CharData *ch, char *argument, int cmd, int subcmd);
+
+#endif //BYLINS_SRC_CMD_DO_AFFECTS_H_
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
