+++ src/cmd/do_mode.h
@@ -1,0 +1,14 @@
+//
+// Created by Sventovit on 07.09.2024.
+//
+
+#ifndef BYLINS_SRC_CMD_DO_MODE_H_
+#define BYLINS_SRC_CMD_DO_MODE_H_
+
+class CharData;
+
+void do_mode(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/);
+
+#endif //BYLINS_SRC_CMD_DO_MODE_H_
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
