+++ src/cmd/do_toggle.h
@@ -1,0 +1,11 @@
+//
+// Created by Sventovit on 07.09.2024.
+//
+
+#ifndef BYLINS_SRC_CMD_DO_TOGGLE_H_
+#define BYLINS_SRC_CMD_DO_TOGGLE_H_
+
+class CharData;
+void do_toggle(CharData *ch, char * /*argument*/, int/* cmd*/, int/* subcmd*/);
+
+#endif //BYLINS_SRC_CMD_DO_TOGGLE_H_
